+++ src/lib/converter/factories/signature.ts
@@ -286,10 +286,10 @@ function convertTypeParameters(
 
         // There's no way to determine directly from a ts.TypeParameter what it's variance modifiers are
         // so unfortunately we have to go back to the node for this...
-        const variance = getVariance(
-            param.getSymbol()?.declarations?.find(ts.isTypeParameterDeclaration)
-                ?.modifiers
-        );
+        const declaration = param
+            .getSymbol()
+            ?.declarations?.find(ts.isTypeParameterDeclaration);
+        const variance = getVariance(declaration?.modifiers);
 
         const paramRefl = new TypeParameterReflection(
             param.symbol.name,
@@ -304,6 +304,15 @@ function convertTypeParameters(
         paramRefl.default = defaultT
             ? context.converter.convertType(paramCtx, defaultT)
             : void 0;
+
+        // No way to determine this from the type parameter itself, need to go back to the declaration
+        if (
+            declaration?.modifiers?.some(
+                (m) => m.kind === ts.SyntaxKind.ConstKeyword
+            )
+        ) {
+            paramRefl.flags.setFlag(ReflectionFlag.Const, true);
+        }
 
         context.registerReflection(paramRefl, param.getSymbol());
         context.trigger(ConverterEvents.CREATE_TYPE_PARAMETER, paramRefl);
@@ -337,6 +346,10 @@ export function createTypeParamReflection(
     paramRefl.default = param.default
         ? context.converter.convertType(paramScope, param.default)
         : void 0;
+    if (param.modifiers?.some((m) => m.kind === ts.SyntaxKind.ConstKeyword)) {
+        paramRefl.flags.setFlag(ReflectionFlag.Const, true);
+    }
+
     context.registerReflection(paramRefl, param.symbol);
 
     if (ts.isJSDocTemplateTag(param.parent)) {