+++ src/lib/output/themes/lib.tsx
@@ -109,6 +109,7 @@ export function renderTypeParametersSignature(
                     <span class="tsd-signature-symbol">{"<"}</span>
                     {join(<span class="tsd-signature-symbol">{", "}</span>, typeParameters, (item) => (
                         <>
+                            {item.flags.isConst && "const "}
                             {item.varianceModifier ? `${item.varianceModifier} ` : ""}
                             <span class="tsd-signature-type" data-tsd-kind={ReflectionKind.singularString(item.kind)}>
                                 {item.name}