+++ src/lib/types/ts-internal/index.d.ts
@@ -10,8 +10,15 @@ declare module "typescript" {
     }
 
     interface Symbol {
+        // TS before 5.0
         // https://github.com/microsoft/TypeScript/blob/v4.1.5/src/compiler/types.ts#L4734-L4737
         checkFlags?: CheckFlags;
+
+        // TS 5.0
+        // https://github.com/microsoft/TypeScript/blob/5.0.2/src/compiler/types.ts#L5891-L5898
+        links?: {
+            checkFlags: CheckFlags;
+        };
 
         // https://github.com/microsoft/TypeScript/blob/v4.7.4/src/compiler/types.ts#L4941
         // https://github.com/microsoft/TypeScript/issues/38344