+++ src/lib/utils/options/declaration.ts
@@ -167,6 +167,7 @@ export interface TypeDocOptionMap {
 
     // Validation
     treatWarningsAsErrors: boolean;
+    treatValidationWarningsAsErrors: boolean;
     intentionallyNotExported: string[];
     validation: ValidationOptions;
     requiredToBeDocumented: ReflectionKind.KindString[];