+++ src/test/converter2/behavior/overloadTags.js
@@ -1,0 +1,18 @@
+/**
+ * @overload
+ * @param {string} value first docs
+ * @return {void}
+ */
+
+/**
+ * @overload
+ * @param {number} value second docs
+ * @param {number} [maximumFractionDigits] second docs
+ * @return {void}
+ */
+
+/**
+ * @param {string | number} value impl docs
+ * @param {number} [maximumFractionDigits] impl docs
+ */
+export function printValue(value, maximumFractionDigits) {}