+++ src/test/converter2/issues/gh2200.ts
@@ -1,0 +1,5 @@
+declare function buildObj<T>(x: T): {
+    [K in keyof T]?: 1;
+};
+
+export const Test = buildObj({ x: 1 });